Only send unix credentials on linux
