+++ transport_darwin.go
@@ -1,0 +1,7 @@
+package dbus
+
+func (t *unixTransport) SendNullByte() error {
+	_, err := t.Write([]byte{0})
+	return err
+}
+
